STYLE: Replace `static_cast<unsigned int>` on integers with `U` suffixes
Following C++ Core Guidelines, Jun 14, 2026, Avoid casts, https://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#res-casts

Uppercase suffix `U` appears (somewhat?) preferable to lowercase `u`, according to https://clang.llvm.org/extra/clang-tidy/checks/readability/uppercase-literal-suffix.html

+++ Components/Optimizers/AdaptiveStochasticGradientDescent/elxAdaptiveStochasticGradientDescent.hxx
@@ -172,7 +172,7 @@ AdaptiveStochasticGradientDescent<TElastix>::BeforeEachResolution()
      * This is a rather crude rule of thumb, which seems to work in practice.
      */
     this->m_NumberOfJacobianMeasurements =
-      std::max(static_cast<unsigned int>(1000), static_cast<unsigned int>(numberOfParameters));
+      std::max(1000U, static_cast<unsigned int>(numberOfParameters));
     configuration.ReadParameter(
       this->m_NumberOfJacobianMeasurements, "NumberOfJacobianMeasurements", this->GetComponentLabel(), level, 0);
